requirements.txt
openai>=1.0.0
pyyaml>=6.0
jinja2>=3.1.0
pdfkit>=1.0.0
click>=8.1.0
python-dotenv>=1.0.0